+++ src/views/mod.rs
@@ -2,6 +2,7 @@ mod blog_list;
 mod blog_post;
 mod design_system;
 mod landing;
+mod music;
 mod projects_list;
 mod resume;
 
@@ -9,5 +10,6 @@ pub use blog_list::BlogList;
 pub use blog_post::BlogPost;
 pub use design_system::DesignSystem;
 pub use landing::Landing;
+pub use music::Music;
 pub use projects_list::ProjectsList;
 pub use resume::ResumePage;